+++ InvalidMoveException.java
@@ -1,5 +1,3 @@
-package ChessPiece;
-
 public class InvalidMoveException extends Exception
 {
     public InvalidMoveException(String message)
